+++ app/Http/Controllers/Api/SettingController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
+use App\Http\Resources\SettingResource;
 use App\Models\Setting;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
@@ -36,9 +37,26 @@ class SettingController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show()
     {
-        //
+        $setting = new SettingResource(Setting::first());
+
+        if ($setting) {
+            return response()->json(
+                [
+                    'message' => 'Setting Information',
+                    'data' => $setting,
+                ],
+                200
+            );
+        } else {
+            return response()->json(
+                [
+                    'message' => 'Setting Failed',
+                ],
+                400
+            );
+        }
     }
 
     /**
@@ -53,8 +71,7 @@ class SettingController extends Controller
 
         $validator = Validator::make($request->all(), [
             'name' => 'required',
-            'logo' => 'required|image|mimes:jpeg,png,jpg',
-            'favicon' => 'required|image|mimes:jpeg,png,jpg',
+            'short_description' => 'required',
             'email' => 'required|email',
             'mobile' => 'required|min:11|max:11',
             'address' => 'required',
@@ -71,46 +88,48 @@ class SettingController extends Controller
 
         $setting = Setting::find($id);
 
-        if ($request->hasFile('logo')) {
+        $data = [
+            'name' => $request->name,
+            'short_description' => $request->short_description,
+            'email' => $request->email,
+            'mobile' => $request->mobile,
+            'address' => $request->address,
+            'fb_link' => $request->fb_link,
+            'twitter_link' => $request->twitter_link,
+            'linekdin_link' => $request->linekdin_link,
+        ];
 
-            $destination = public_path($setting->logo);
-            if (file_exists($destination)) {
+        if ($request->hasFile('logo')) {
+            $destination = public_path($setting->image);
+
+            if ($setting->image && file_exists($destination)) {
                 unlink($destination);
             }
-            
-            $image = $request->file('logo');
-            $extension = $image->extension();
-            $name = time().'.'.$extension;
-            $image->move(public_path('/upload/logo/'), $name);
-            $logo_path = 'upload/logo/'.$name;
+
+            $logo = $request->file('logo');
+            $extension = $logo->extension();
+            $name =  'logo-' . time() . '.' . $extension;
+            $logo->move(public_path('/upload/logo/'), $name);
+            $logo_path = 'upload/logo/' . $name;
+            $data['logo'] = $logo_path;
         }
         if ($request->hasFile('favicon')) {
-
             $destination = public_path($setting->favicon);
-            if (file_exists($destination)) {
+
+            if ($setting->favicon && file_exists($destination)) {
                 unlink($destination);
             }
 
-            $image = $request->file('favicon');
-            $extension = $image->extension();
-            $name = time().'.'.$extension;
-            $image->move(public_path('/upload/favicon/'), $name);
-            $favicon_path = 'upload/favicon/'.$name;
+            $favicon = $request->file('favicon');
+            $extension = $favicon->extension();
+            $name = 'favicon-' . time() . '.' . $extension;
+            $favicon->move(public_path('/upload/logo/'), $name);
+            $favicon_path = 'upload/logo/' . $name;
+            $data['favicon'] = $favicon_path;
         }
 
-        $setting->name = $request->name;
-        $setting->email = $request->email;
-        $setting->mobile = $request->mobile;
-        $setting->address = $request->address;
-        $setting->fb_link = $request->fb_link;
-        $setting->twitter_link = $request->twitter_link;
-        $setting->linekdin_link = $request->linekdin_link;
-        $setting->logo = $logo_path;
-        $setting->favicon = $favicon_path;
+        $setting->update($data);
 
-
-        $setting->save();
-        
         return response()->json([
             'message' => 'Setting Updated Successfully',
             'data' =>  $setting,
